PROJECT OBJECTIVE:
COVID19 also known as Corona Virus as been in existence since 2020 and its has since then spur on widespread fear and paranoia because it is highly contagious. This pandemic is felt across the world in the Economic System, Health System, Educational Sector and our way of life generally.  However, the purpose of this project is to show a data driven decision on the growth or increase from the reported cases of sum of confirmed, sum of deaths and rate.